+++ pkg/key/key.go
@@ -29,8 +29,6 @@ const (
 	WriteClusterNamespaceAppsRole        = "write-in-cluster-ns"
 	WriteOrganizationsPermissionsName    = "write-organizations"
 	WriteFluxResourcesPermissionsName    = "write-flux-resources"
-	WriteClustersPermissionsName         = "write-clusters"
-	WriteNodePoolsPermissionsName        = "write-nodepools"
 	WriteClientCertsPermissionsName      = "write-client-certificates"
 	WriteSilencesPermissionsName         = "write-silences"
 )
@@ -135,14 +133,6 @@ func WriteOrganizationsCustomerGroupClusterRoleBindingName() string {
 
 func WriteFluxResourcesAutomationSARoleBindingName() string {
 	return fmt.Sprintf("%s-customer-sa", WriteFluxResourcesPermissionsName)
-}
-
-func WriteClustersAutomationSARoleBindingName() string {
-	return fmt.Sprintf("%s-customer-sa", WriteClustersPermissionsName)
-}
-
-func WriteNodePoolsAutomationSARoleBindingName() string {
-	return fmt.Sprintf("%s-customer-sa", WriteNodePoolsPermissionsName)
 }
 
 func WriteClientCertsAutomationSARoleBindingName() string {